requirements.txt
torch>=1.9.0
torchvision>=0.10.0
datasets>=2.0.0
Pillow>=8.0.0
numpy>=1.19.0
tqdm>=4.62.0
matplotlib>=3.4.0 